+++ src/index.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export const VERSION = "0.2.3"; // x-release-please-version
+export const VERSION = "0.2.6"; // x-release-please-version
 
 import * as path from "node:path";